Add rq
This adds rq to the pipfile and also specifies the versions of flask
that the project relies on.

There was some consideration of NOT specifying versions, since they
appear in the lock, but ultimately it seems like a good practice to be
explicit about which version of a package the project expects.

Issue #6