Update documentation
This creates a new README which is more specific to the SimpleBook
project, particularly for development of it.

We're going to keep the original documentation as well since it is
probably still useful.  At some point if it becomes obsolete with the
SimpleBook documentation we can remove that as well

While updating the documentation, there is a slight code re-organization
which puts the `api` code into a `services` directory.  The intent is
that by adding this folder it makes it more clear that the API is
a service that the SimpleBook plugin relies on.  There's a very small
benefit of also giving a good spot for future related services.

Issue #5