Add new documentation
This creates a new README which is more specific to the SimpleBook
project, particularly for development of it.

We're going to keep the original documentation as well since it is
probably still useful.  At some point if it becomes obsolete with the
SimpleBook documentation we can remove that as well

Issue #5