Add node MediaWiki authentication
Authentication is currently done outside of the node application and
this tool expects a cookie jar to be passed.  This actually does make
sense as an option from a generic library perspective (e.g. if mw2pdf
was url2pdf).

This change is one step in adding optional parameters so that
credentials and MediaWiki api information can be provided and the tool
will authenticate in its own session.

This does not mean we will remove the ability to accept externally
defined cookies.  We may decide to remove this logic and put it in the
SimpleBook API instead, at which point we might consider doing that
generic renaming of the package.

Issue #11