Support SimpleBook authentication
Authentication is being handled similar to the way the original
Collection plugin worked: the MediaWiki instance calls the pdf rendering
service, passes credentials to use when rendering, and those credentials
are used.

We're making two changes:

1. Instead of using string manipulation to pass those credentials, we
   are now going to have those credentials passed as individual fields.

2. There is now support for "passthrough" data, which is a json string
   that gets converted to request parameters on individual render calls.

The goal of the first change is just to make the format for
authentication data more explicit.

The goal of the second change is to allow individual mediawiki instances
to have more control; for instance, it might pass contextual information
around what pages / data should be rendered based on what the logged in
user has access to.

SimpleBook itself doesn't need to know or care about the nature of the
passthrough data, it just commits to forwarding it along.

Issue #9