Add validation to passthroughParameters
Instead of passively failing if the passthroughParameters does not
contain a valid JSON string we can be explicit and validate using Joi.

This also makes sure that the passthroughParameters is an object with
primitive values for each key.  This puts the burden of encoding
complex parameters on the application instead of imposing JSON.