Rename extension
There are a lot of things that need to be renamed, but this is a first
step which will at least allow us to install SimpleBook as "SimpleBook"
instead of "Collection".

This is important because we have officially begun to diverge from
Collection.

SimpleBook and Collection should NOT be installed on the same mediawiki
instance at this point, since they share some of the same configuration
variable names.

Issue #25