Rename and reformat 'COPYING' file to 'LICENSE.md'
Move the copyright license file from the older 'COPYING' to the now
somewhat more common 'LICENSE.md', and reformat it from plaintext to
Markdown.  The license itself (GPL-2.0) remains unchanged.

I did this after going looking for the license and not seeing it, not
because it wasn't there but because I'm so used to looking for a file
named "LICENSE" or "LICENSE.md" these days that I totally missed it
under the name "COPYING".  And I grew up using the former!  So if
nowadays I miss it on the first pass, other people probably would too.