Disable chapters
Chapters are not supported by the renderer at this moment, which means
that they break if a user attempts to create a book with chapters.  In
the long run we would like to support chapters, but in the mean time  we
will disable the UX to prevent a broken user experience.

Issue #37