Support missing parameters
The python was accessing form parameters through direct access, which
would throw KeyErrors if that form field was not set.  This was causing
book failures in situations where a username / password wasn't being
set.

Using `.get` does not throw an error, and instead returns a None value.

The rest of the code already supported the possibility of having no
authentication credentials.

Issue #50